+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 next
 ----
 
+- Require registration in ``mpl.rcParams["figure.hooks"]`` for the
+  :envvar:`MPLCURSORS` environment variable.
 - Support new-style (Matplotlib≥3.8) contour plots.
 
 0.6